Minor changes so dzil and perlcritic play along
The author tests of drill complained that I had code before strictures.
The reason for this was the Plugin::PkgVersion, which put the
declaration of VERSION right after the package declaration and _before_
»use strict«.

There seems to be a pull request for Dist::Zilla that hasn’t been
merged for some time now:
https://github.com/rjbs/dist-zilla/pull/168

Fixed the code so dzil doesn’t complain any more.